+++ repositories/cliente-repository.dart
@@ -1,6 +1,4 @@
-import 'dart:ffi';
-
 import '../models/cliente.dart';
 import 'repository-impl.dart';
 
-class ClienteRepository extends RepositoryImpl<Cliente, Int64> {}
+class ClienteRepository extends RepositoryImpl<Cliente, int> {}
