+++ repositories/produto-repository.dart
@@ -3,4 +3,4 @@ import 'dart:ffi';
 import '../models/produto.dart';
 import 'repository-impl.dart';
 
-class ProdutoRepository extends RepositoryImpl<Produto, Int64> {}
+class ProdutoRepository extends RepositoryImpl<Produto, int> {}
